# internal/keyspan: apply various micro-optimizations