These commits introduce a new microbenchmark `BenchmarkIteratorSeekNoRangeKeys`
and apply a few micro-optimizations that help move the needle on cockroachdb/cockroach#82559.

```
name                                    old time/op    new time/op    delta
IteratorSeekNoRangeKeys/batch=false-10    5.09µs ± 1%    4.28µs ± 1%  -16.03%  (p=0.000 n=10+19)
IteratorSeekNoRangeKeys/batch=true-10     9.85µs ± 1%    8.78µs ± 1%  -10.87%  (p=0.000 n=9+20)
```